% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super_families.get_hypotheses_at_level.R
\name{super_families.get_hypotheses_at_level}
\alias{super_families.get_hypotheses_at_level}
\title{Method for Super Family Hypotheses Trees}
\usage{
super_families.get_hypotheses_at_level(super_families, level = 1)
}
\arguments{
\item{super_families}{Hypotheses tree of type \code{data.tree::Node}.}

\item{level}{Level of the hypotheses within the tree.}
}
\value{
A list of nodes containing all hypotheses at the given level.
}
\description{
This function returns the hypotheses at a given level.
}
\details{
The function \code{super_families.get_hypotheses_at_level} translates
\code{nodes$level} of all nodes in \code{super_families} to the corresponding
hypotheses level. For all nodes the attribute "hypothesis_level" is added or
updated.
}
\examples{
super_families <- Node$new("H31")
super_families$AddChild("H21")$AddChild("H11")$AddSibling("H12")
super_families$AddChild("H13")

super_families.get_hypotheses_at_level(super_families)

print(super_families, "level", "hypothesis_level")
}
