% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super_families.test.R
\name{super_families.test}
\alias{super_families.test}
\title{Test of a Super Family Hypotheses Tree}
\usage{
super_families.test(super_families, kappa = 0.001)
}
\arguments{
\item{super_families}{Hypotheses tree of type \code{data.tree::Node}.}

\item{kappa}{The parameter \code{kappa} describes the percentage of signals
(small p-values) needed for a level-1 hypotheses to be rejected, i.e., to
be not considered as noise only. See the parameter \code{kappa} in
\code{\link{hierasymptkappa}}.}
}
\value{
The hypotheses tree super_families with added or updated attribute
  "test_results" for level-1 hypotheses (logical vector) and attribute "rejected" for
  all nodes. (TRUE or FALSE)
}
\description{
Tests all hypotheses nodes contained in a super family hypotheses tree.
}
\details{
The nodes in \code{super_families} are tested recursively. For testing the
level-1 hypotheses the attribute "p_values" assigned with a vector of
p-values is required. The function \code{\link{hierasymptkappa}} is applied
on the the list of p-value vectors. For level larger than one a hypothesis is
rejected if and only if at least one child hypothesis is rejected.
}
\examples{
super_families <- Node$new("H31")
super_families$AddChild("H21")$AddChild("H11")$AddSibling("H12")
super_families$AddChild("H13")

set.seed(1)
for (H in super_families.get_hypotheses_at_level(super_families, level = 1)) {
  p_values <- runif(10)
  ind <- sample(1:10, sample(c(0,1)))
  p_values[ind] <- 0
  Set(list(H), p_values = list(p_values))
}
rm(ind, p_values)
super_families.test(super_families)
Get(Traverse(super_families), "p_values")
Get(Traverse(super_families), "test_results")

print(super_families, "rejected")
}
